+++ src/main/java/com/server/epigram/common/EmotionType.java
@@ -1,0 +1,5 @@
+package com.server.epigram.common;
+
+public enum EmotionType {
+    MOVED, HAPPY, WORRIED, SAD, ANGRY
+}
