+++ src/main/java/com/server/epigram/common/OauthType.java
@@ -1,0 +1,5 @@
+package com.server.epigram.common;
+
+public enum OauthType {
+    KAKAO, GOOGLE
+}
