+++ src/main/java/com/server/epigram/common/TokenType.java
@@ -1,0 +1,5 @@
+package com.server.epigram.common;
+
+public enum TokenType {
+    OAUTH, LOCAL
+}
