--- src/main/java/com/server/epigram/global/common/EmotionType.java
@@ -1,5 +1,0 @@
-package com.server.epigram.global.common;
-
-public enum EmotionType {
-  MOVED, HAPPY, WORRIED, SAD, ANGRY
-}
