--- src/main/java/com/server/epigram/global/common/TokenType.java
@@ -1,5 +1,0 @@
-package com.server.epigram.global.common;
-
-public enum TokenType {
-  OAUTH, LOCAL
-}
